pyproject.toml
[tool.poetry]
name = "checkbot"
version = "0.1.0"
description = "Discord bot to time track employees"
authors = ["Hudson Bruno <hb.natal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.0"
jsonschema = {extras = ["format"], version = "^4.19.0"}
rfc3339-validator = "^0.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
ruff = "^0.0.287"
watchdog = {extras = ["watchmedo"], version = "^3.0.0"}

[tool.isort]
profile = 'black'

[tool.black]
preview = true

[tool.coverage.paths]
source = ["src"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
